requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
dj-rest-auth==2.1.9
Django==3.2.18
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
oauthlib==3.2.2
Pillow==9.5.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
